Add condition for unlink only when not withdrawing test data

+++ tests3/acceptance/Backend/TestMaintenanceCest.php
@@ -88,7 +88,13 @@ class TestMaintenanceCest
 		$I->clickAndWait(MaintenancePage::$saveTablesButton, 10);
 
 		$I->assertTrue(file_exists($downloadPath));
-		unlink($downloadPath);
+
+		$withdraw = getenv('BW_TEST_WITHDRAW');
+
+		if ($withdraw == false)
+		{
+			unlink($downloadPath);
+		}
 	}
 
 	/**
@@ -135,7 +141,13 @@ class TestMaintenanceCest
 		$I->clickAndWait(MaintenancePage::$saveTablesButton, 10);
 
 		$I->assertTrue(file_exists($downloadPath));
-		unlink($downloadPath);
+
+		$withdraw = getenv('BW_TEST_WITHDRAW');
+
+		if ($withdraw == false)
+		{
+			unlink($downloadPath);
+		}
 
 		$I->setManifestOption('com_bwpostman', 'compress_backup', '1');
 	}
